bar: reorganize files, update build system.
Reorganize includes, declarations and definitions:
 - Move standalone functions to bar_tools.cpp
 - Move class methods to bar_classes.cc
 - Use bar.h include in all files, to allow compilation of each file
 individually
 - Fix global vars:
   - Move non-global variables to main() local scope
   - Use extern declaration for global variables

Update build system:
 - Add all files to meson.build instead of including them into the main file

Small indentation fix.
