+++ bar/bar_tools.cc
@@ -6,6 +6,8 @@
  * License: GPL3
  * */
 
+#include "bar.h"
+
 /*
  * Returns config dir
  * */
@@ -198,31 +200,15 @@ Gtk::Image* app_image(std::string icon) {
     return image;
 }
 
-/*
- * Argument parser
- * Credits for this cool class go to iain at https://stackoverflow.com/a/868894
- * */
-class InputParser{
-    public:
-        InputParser (int &argc, char **argv){
-            for (int i=1; i < argc; ++i)
-                this->tokens.push_back(std::string(argv[i]));
-        }
-        /// @author iain
-        const std::string& getCmdOption(const std::string &option) const{
-            std::vector<std::string>::const_iterator itr;
-            itr =  std::find(this->tokens.begin(), this->tokens.end(), option);
-            if (itr != this->tokens.end() && ++itr != this->tokens.end()){
-                return *itr;
-            }
-            static const std::string empty_string("");
-            return empty_string;
-        }
-        /// @author iain
-        bool cmdOptionExists(const std::string &option) const{
-            return std::find(this->tokens.begin(), this->tokens.end(), option)
-                   != this->tokens.end();
-        }
-    private:
-        std::vector <std::string> tokens;
-};
+void on_button_clicked(std::string cmd) {
+    cmd = cmd + " &";
+    const char *command = cmd.c_str();
+    std::system(command);
+
+    Gtk::Main::quit();
+}
+
+gboolean on_window_clicked(GdkEventButton *event) {
+    Gtk::Main::quit();
+    return true;
+}
